Add logic to create cart (if it is not present) when creating cartItem. Change few method signatures to implement that new logic
