+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/repository/CartRepository.java
@@ -1,9 +1,10 @@
 package com.avdo.spring.app.repository;
 
 import com.avdo.spring.app.service.domain.model.Cart;
+import com.avdo.spring.app.service.domain.model.User;
 
 public interface CartRepository {
 
     Cart findByUserEntityId(Long userId);
-    Cart createOrFindCart();
+    Cart createCart(User user);
 }
