+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/service/CartService.java
@@ -1,9 +1,10 @@
 package com.avdo.spring.app.service;
 
 import com.avdo.spring.app.service.domain.model.Cart;
+import com.avdo.spring.app.service.domain.model.User;
 
 public interface CartService {
 
-    Cart createCart();
+    Cart createCart(User user);
     Cart findByUserId(Long id);
 }
