+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/service/impl/CartServiceImpl.java
@@ -3,6 +3,7 @@ package com.avdo.spring.app.service.impl;
 import com.avdo.spring.app.repository.CartRepository;
 import com.avdo.spring.app.service.CartService;
 import com.avdo.spring.app.service.domain.model.Cart;
+import com.avdo.spring.app.service.domain.model.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -17,8 +18,8 @@ public class CartServiceImpl implements CartService {
     }
 
     @Override
-    public Cart createCart() {
-        return cartRepository.createOrFindCart();
+    public Cart createCart(User user) {
+        return cartRepository.createCart(user);
     }
 
     @Override
